chore(llm): crypto polyfill rework
