+++ .changeset/sharp-keys-explode.md
@@ -1,0 +1,5 @@
+---
+"live-mobile": minor
+---
+
+Implement crypto polyfill with react-native-quick-crypto for better implementation and better performance.
