+++ apps/ledger-live-mobile/.unimportedrc.json
@@ -34,6 +34,7 @@
     "react-native-randombytes",
     "react-native-tcp",
     "react-native-tcp-socket",
-    "react-native-udp"
+    "react-native-udp",
+    "safe-buffer"
   ]
 }
