+++ apps/ledger-live-mobile/package.json
@@ -83,6 +83,7 @@
     "@ledgerhq/ethereum-provider": "workspace:^",
     "@ledgerhq/hw-transport": "workspace:^",
     "@ledgerhq/hw-transport-http": "workspace:^",
+    "@ledgerhq/hw-trustchain": "workspace:^",
     "@ledgerhq/icons-ui": "workspace:^",
     "@ledgerhq/live-common": "workspace:^",
     "@ledgerhq/live-config": "workspace:^",
@@ -101,7 +102,6 @@
     "@ledgerhq/types-cryptoassets": "workspace:^",
     "@ledgerhq/types-devices": "workspace:^",
     "@ledgerhq/types-live": "workspace:^",
-    "@ledgerhq/hw-trustchain": "workspace:^",
     "@likashefqet/react-native-image-zoom": "^1.3.0",
     "@react-native-async-storage/async-storage": "^1.19.2",
     "@react-native-clipboard/clipboard": "^1.11.2",
@@ -172,7 +172,7 @@
     "react-native-fast-crypto": "^2.2.0",
     "react-native-fast-image": "^8.5.11",
     "react-native-gesture-handler": "^2.9.0",
-    "react-native-get-random-values": "^1.11.0",
+    "react-native-get-random-values": "1",
     "react-native-haptic-feedback": "^2.0.3",
     "react-native-image-crop-tools": "^1.6.4",
     "react-native-image-picker": "^5.6.1",
@@ -185,6 +185,7 @@
     "react-native-navigation-bar-color": "^2.0.1",
     "react-native-pager-view": "^6.0.1",
     "react-native-qrcode-svg": "6.1.1",
+    "react-native-quick-crypto": "^0.7.1",
     "react-native-randombytes": "^3.6.1",
     "react-native-reanimated": "3.8.1",
     "react-native-restart": "^0.0.24",
@@ -210,6 +211,7 @@
     "reselect": "4.0.0",
     "rn-fetch-blob": "^0.12.0",
     "rxjs": "^7.8.1",
+    "safe-buffer": "^5.2.1",
     "semver": "^7.3.7",
     "storyly-react-native": "2.0.1",
     "styled-components": "^5.3.3",
