+++ apps/ledger-live-mobile/src/polyfill-crypto.ts
@@ -1,0 +1,5 @@
+// + "crypto" node polyfill
+// + global.crypto web polyfill
+
+import { install } from "react-native-quick-crypto";
+install();
