+++ libs/coin-modules/coin-evm/package.json
@@ -77,7 +77,6 @@
     "imurmurhash": "^0.1.4",
     "invariant": "^2.2.2",
     "lodash": "^4.17.21",
-    "react-native-get-random-values": "^1.8.0",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
