+++ libs/coin-modules/coin-evm/src/api/node/rpc.native.ts
@@ -1,7 +1,5 @@
 /* istanbul ignore file: pure exports, test unecessary */
 
-/** ⚠️ keep this order of import. @see https://docs.ethers.io/v5/cookbook/react-native/#cookbook-reactnative ⚠️ */
-import "react-native-get-random-values";
 import "@ethersproject/shims";
 export * from "./rpc.common";
 export * as default from "./rpc.common";
